Panel Level manager On progress

+++ Assets/Scripts/Database/SaveData.cs
@@ -8,7 +8,7 @@ public class SaveData
 
     public static float GetPlayerProgress(string key) => PlayerPrefs.GetFloat(key);
 
-    public static void SetSetting(string key, int value) => PlayerPrefs.SetInt(key, value);
+    public static void SetSetting(string key, float value) => PlayerPrefs.SetFloat(key, value);
 
-    public static int GetSettings(string key) => PlayerPrefs.GetInt(key);
+    public static float GetSettings(string key) => PlayerPrefs.GetFloat(key);
 }
